server/requirements.txt
aiofiles==0.8.0
anyio==3.5.0
asgiref==3.5.1
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
fastapi==0.76.0
fastapi-login==1.8.0
Flask==2.1.2
h11==0.13.0
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.8.0
MarkupSafe==2.1.1
passlib==1.7.4
pydantic==1.9.0
PyExecJS==1.5.1
PyJWT==2.3.0
python-multipart==0.0.5
requests==2.27.1
six==1.16.0
sniffio==1.2.0
starlette==0.18.0
typing_extensions==4.2.0
urllib3==1.26.9
uvicorn==0.17.6
Werkzeug==2.1.2
